feat: adicionando middleware de validação na atualização de menbro da família
